% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{add_author_folder}
\alias{add_author_folder}
\title{Add author folder}
\usage{
add_author_folder(initials, folder, con, schema = "platform")
}
\arguments{
\item{initials}{keep it short, also unique!}

\item{folder}{can be left empty}

\item{con}{connection to the database}

\item{schema}{schema name}
}
\value{
number of rows inserted
}
\description{
Add the folder to an author already in the database.
}
