% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{check_data_df}
\alias{check_data_df}
\title{Check data to be parsed}
\usage{
check_data_df(df, con)
}
\arguments{
\item{df}{data with time series data}

\item{con}{connection to the database}
}
\value{
TRUE if it works
}
\description{
Check data to be parsed
}
