% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{check_data_xlsx}
\alias{check_data_xlsx}
\title{Check data file to be parsed}
\usage{
check_data_xlsx(filename, codes)
}
\arguments{
\item{filename}{Excel file with time series data}

\item{codes}{character vector of codes}

\item{con}{connection to the database}
}
\value{
TRUE if it works
}
\description{
checks all the sheets with the data are
ok with \link[UMARfetchR]{check_data_df}
}
