% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{check_structure_df}
\alias{check_structure_df}
\title{Check structure data to be parsed}
\usage{
check_structure_df(df, interval)
}
\arguments{
\item{df}{dataframe from one of the structure worksheets}

\item{interval}{name of the sheet - which is the one character interval code e.g. "M" or "Q"}
}
\value{
TRUE if passes all checks
}
\description{
This function checks if the input data that is read from an excel file is
in the proper format to be parsed by the parser functions and gives informative
errors if not.
}
