% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{check_structure_wb}
\alias{check_structure_wb}
\title{Check structure workbook to be parsed}
\usage{
check_structure_wb(filename)
}
\arguments{
\item{filename}{path to excel file}
}
\value{
TRUE if all checks pass
}
\description{
Takes the excel file with the data to be parsed and checks the inputs are all
OK.
}
