% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{compute_series_codes}
\alias{compute_series_codes}
\title{Compute the series codes for the new series}
\usage{
compute_series_codes(df)
}
\arguments{
\item{df}{a dataframe with the table code, dimension levels code, interval and}
}
\value{
df with a complete series_code column
}
\description{
To be run after \link[UMARfetchR]{compute_table_codes} is already run so
all the series have a legit table code. Computes the series code, which will
be unique, because of the checks and how the table codes are computed.
}
