% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{compute_series_names}
\alias{compute_series_names}
\title{Compute the series codes for the new series}
\usage{
compute_series_names(df)
}
\arguments{
\item{df}{a dataframe with dimension_levels_text}
}
\value{
df with a complete series_code column
}
\description{
This isn't really a compute funciton, because it just copies the values, but
kinda fits with the other funcitons to name it this way. Anyway, in case there
are no series name values, this funciton copies the ones from the
dimension_levels_text field.
}
