% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{compute_table_codes}
\alias{compute_table_codes}
\title{Compute the table codes for new series}
\usage{
compute_table_codes(df, con)
}
\arguments{
\item{df}{dataframe from structure template}

\item{con}{connection to the database.}
}
\value{
dataframe with a complete table_code column
}
\description{
Computes the new table codes for the tables of the series to be imported.
This uses the author field to create the code and checks which ones are
already in the database and increments from there.
}
\details{
If the table already exists, need to cover this eventuality
}
