% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{create_data_template_excel}
\alias{create_data_template_excel}
\title{Prepare excel template for data file}
\usage{
create_data_template_excel(path = "", author = "name", overwrite = TRUE)
}
\arguments{
\item{path}{path to file, default empty}

\item{author}{initials of author that becomes part of the filename}

\item{overwrite}{- whether or not to overwrite previous file}
}
\value{
nothing - side effect is saving to an excel file.
}
\description{
Prepares an empty excel file for the data. Really only names the three sheets
}
