% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{create_structure_template_excel}
\alias{create_structure_template_excel}
\title{Prepare excel template for structure file}
\usage{
create_structure_template_excel(path = "", author = "name", overwrite = TRUE)
}
\arguments{
\item{author}{initials of author that becomes part of the filename}

\item{overwrite}{- whether or not to overwrite previous file}
}
\value{
nothing - side effect is saving to an excel file.
}
\description{
Prepares an empty excel file with the correct columns for a structure file.
}
