% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{excel_date_to_r_date}
\alias{excel_date_to_r_date}
\title{Function to convert Excel serial date numbers to R Date objects}
\usage{
excel_date_to_r_date(excel_date)
}
\arguments{
\item{numeric}{excel number}
}
\description{
Function to convert Excel serial date numbers to R Date objects
}
\keyword{internal}
