% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_initials}
\alias{get_initials}
\title{Get the initials from a name}
\usage{
get_initials(name)
}
\arguments{
\item{name}{name}
}
\description{
This function takes a name and returns the initials. It treats both "-" and " "
as separators. This function is for internal use and is not exported.
}
