% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_interval_from_period}
\alias{get_interval_from_period}
\title{Get interval from the period value}
\usage{
get_interval_from_period(period)
}
\arguments{
\item{period}{one of the possible period values}
}
\value{
"M", "A" or "Q" - or error if none work
}
\description{
Get interval from the period value
}
