% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_table_id_with_same_name}
\alias{get_table_id_with_same_name}
\title{Get function to check if the author already has a table with that name}
\usage{
get_table_id_with_same_name(auth, table_name, con)
}
\arguments{
\item{auth}{author initials}

\item{table_name}{table name}

\item{con}{connection to the database}
}
\description{
Checks the author's tables if the same name already exists and uses
that code, otherwise returns NA
}
