% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_data_points}
\alias{insert_data_points}
\title{Insert datapoints into data_point table}
\usage{
insert_data_points(data, con, schema = "platform")
}
\arguments{
\item{data}{dataframes with the data_points
output of the parser}

\item{con}{connection to database}

\item{schema}{is the database schema}
}
\value{
nothing, just some printing along the way
}
\description{
So, the function extracts and preps the data with
and writes it to a temporary table in the database.
}
\details{
It inserts any new periods into the period table,
adds the data points to the data point table.
}
