% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_author}
\alias{insert_new_author}
\title{Insert new author}
\usage{
insert_new_author(name, initials, email, folder = NA, con, schema = "platform")
}
\arguments{
\item{name}{name}

\item{initials}{keep it short, also unique!}

\item{email}{email}

\item{folder}{can be left empty}

\item{con}{connection to the database}

\item{schema}{schema name}
}
\value{
number of rows inserted
}
\description{
Used to insert a new author into the database.
}
