% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_category_table}
\alias{insert_new_category_table}
\title{Insert new category table}
\usage{
insert_new_category_table(df, con, schema = "platform")
}
\arguments{
\item{df}{dataframe with author,and  table code columns}

\item{con}{connection to the database.}

\item{schema}{db schema, defaults to platfrom}
}
\value{
nothing
}
\description{
Used to insert the new UMAR category_table into the database.
}
