% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_dimension_levels}
\alias{insert_new_dimension_levels}
\title{Insert new dimension_levels table}
\usage{
insert_new_dimension_levels(df, con, schema = "platform")
}
\arguments{
\item{df}{dataframe table_code and dimensions, dimension_level_text and
dimension_level_code columns.}

\item{con}{connection to the database.}

\item{schema}{db schema, defaults to platfrom}
}
\value{
nothing
}
\description{
Used to insert the new dimension levels into the database. The input
dataframe must have passed through \link[UMARfetchR]{compute_table_codes} so that
the table_codes and table_names are aligned. And all the checks of the parser
of course.
}
