% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_series_levels}
\alias{insert_new_series_levels}
\title{Insert new series levels table}
\usage{
insert_new_series_levels(df, con, schema = "platform")
}
\arguments{
\item{df}{dataframe with table_code, dimensions and dimension_levels_code and
series_code columns}

\item{con}{connection to the database.}

\item{schema}{db schema, defaults to platfrom}
}
\value{
nothing
}
\description{
Used to insert the new series levels into the database. The input
dataframe must have passed through \link[UMARfetchR]{compute_series_codes}. And all the checks of the parser
of course.
}
