% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_table}
\alias{insert_new_table}
\title{Insert new table}
\usage{
insert_new_table(df, con, schema = "platform")
}
\arguments{
\item{df}{dataframe table_code and table_name columns.}

\item{con}{connection to the database.}

\item{schema}{db schema, defaults to platfrom}
}
\value{
nothing
}
\description{
Used to insert the new tables into the database. The input
dataframe must have passed through \link[UMARfetchR]{compute_table_codes} so that
the table_codes and table_names are aligned.
}
