% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_vintage}
\alias{insert_new_vintage}
\title{Insert new vintages}
\usage{
insert_new_vintage(data, con, schema = "platform")
}
\arguments{
\item{data}{dataframes with the data_points output of the parser}

\item{con}{connection to database}

\item{schema}{database schema}
}
\value{
list of tables with counts for each inserted row.
}
\description{
This function inserts a set of new vintages and their corresponding
data points to the database.
}
