% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{message_structure}
\alias{message_structure}
\title{Prepare message describing structure process}
\usage{
message_structure(filename)
}
\arguments{
\item{filename}{path to excel file}
}
\value{
outputs log message
}
\description{
Prepares a log message describing how many new series will be imported
from each worksheet and how many are already there and will be ignored.
}
