% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{parse_structure}
\alias{parse_structure}
\title{Umbrella function for parsing metadata structural file}
\usage{
parse_structure(df, con)
}
\arguments{
\item{df}{dataframe with the structure data, which means at a minumum the columns
source, author, table_name, dimensions, dimension_levels_text, dimension_levels_code,
unit, and interval.}

\item{con}{connection to the database.}
}
\value{
data frame with original columns and added table and series code fields.
}
\description{
Umbrella function that does the following: runs all the checks to see if the
input data is correctly prepared and returns meaningful errors if not. Then
announces the number of new series and old series that will be parsed. Then
computes the values for the table codes and the series codes and if necessary
also the series names.
}
