% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{prep_and_import_structure}
\alias{prep_and_import_structure}
\title{Umbrella function to prepare and import structure metadata}
\usage{
prep_and_import_structure(df, con, schema = "platform")
}
\arguments{
\item{df}{dataframe with the structure data, which means at a minumum the columns
source, author, table_name, dimensions, dimension_levels_text, dimension_levels_code,
unit, and interval.}

\item{con}{connection to the database.}

\item{schema}{schema name, defaults to "platform"}
}
\value{
vector of inserted rows
}
\description{
Umbrella funciton that takes the structure metadata dataframe an author
prepares using the template created with \link[UMARfetchR]{create_structure_template_excel}
and the connection and schema and prepares and imports the following tables into
the database: `table`, `category_table`, `table_dimensions`, `dimension_levels`,
`series` and `series_levels`. Logs number of rows inserted into each table.
}
