% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_category_relationship_table}
\alias{prepare_category_relationship_table}
\title{Prepare table to insert into `category_relationship` table}
\usage{
prepare_category_relationship_table(cat_name, con, schema)
}
\arguments{
\item{cat_name}{character name of category}

\item{con}{connection to the database}

\item{schema}{schema name}
}
\value{
a dataframe with the `id`, `parent_id`, `source_id` for each relationship
betweeh categories
}
\description{
Helper function that manually prepares the category_relationship table for first level only!
Returns table ready to insert into the `category_relationship` table with the db_writing family
of functions from UMARimportR using \link[UMARimportR]{sql_function_call}.
}
\details{
As there is currently only first level categories in UMAR, this function cannot
do others, but this functionality can be added if the need arises.
}
