% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data_table}
\alias{prepare_data_table}
\title{Get and prepare data for import}
\usage{
prepare_data_table(data, selection, con)
}
\arguments{
\item{data}{dataframe with the data_points}

\item{selection}{dataframe with selected series / output from \link[UMARfetchR]{prepare_vintage_table}}

\item{con}{connection to database}
}
\value{
a dataframe with the period_id, value and id values for all the vintages in the table.
}
\description{
Prepares the timeseries data for importing into the database.
}
