% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_dimension_levels_table}
\alias{prepare_dimension_levels_table}
\title{Prepare table to insert into `dimension_levels` table}
\usage{
prepare_dimension_levels_table(df, con, schema = "platform")
}
\arguments{
\item{df}{dataframe with table_code, dimensions and dimension_levels_code}

\item{con}{connection to the database}

\item{schema}{schema name}
}
\value{
a dataframe with the `dimension_id`, `values` and `valueTexts`
columns for this table.
}
\description{
Helper function that manually prepares the dimension_levels for each
table and get their codes and text.
Returns table ready to insert into the `dimension_levels`table with the
db_writing family of functions.
}
