% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_new_author_table}
\alias{prepare_new_author_table}
\title{Prepare data to be inserted into umar author table}
\usage{
prepare_new_author_table(name, initials, email, folder = NA)
}
\arguments{
\item{name}{name}

\item{initials}{keep it short, also unique!}

\item{email}{email}

\item{folder}{can be left empty}
}
\value{
a df with the four columns above.
}
\description{
Prepare data to be inserted into umar author table
}
