% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_periods}
\alias{prepare_periods}
\title{Prepare period values from dates}
\usage{
prepare_periods(data)
}
\arguments{
\item{data}{individual sheet data}
}
\value{
with updated period
}
\description{
Leaves years if the data is annual, and swaps dates for month (2023M02) or
quarterly values (2023Q1)
}
