% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_series_levels_table}
\alias{prepare_series_levels_table}
\title{Prepare table to insert into `series_levels` table}
\usage{
prepare_series_levels_table(df, con, schema = "platform")
}
\arguments{
\item{df}{dataframe with table_code, dimensions and dimension_levels_code and series_code}

\item{con}{connection to the database}

\item{schema}{schema name}
}
\value{
a dataframe with the `series_id`, `tab_dim_id` and `level_value`
columns for this table.
}
\description{
Helper function that manually prepares the series_levels for each
series and get their ids and values (codes).
Returns table ready to insert into the `series_levels`table with the
db_writing family of functions.
}
