% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_series_table}
\alias{prepare_series_table}
\title{Prepare table to insert into `series` table}
\usage{
prepare_series_table(df, con, schema)
}
\arguments{
\item{df}{dataframe with table_code, series_name, series_code, unit, interval}

\item{con}{connection to the database}

\item{schema}{schema name}
}
\value{
a dataframe with the following columns: `name_long`, `code`,
`unit_id`, `table_id` and `interval_id`for each series in the table
well as the same number of rows as there are series
}
\description{
Prepare a series table to import into the series table in the database.
}
