% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_table_table}
\alias{prepare_table_table}
\title{Prepare table to insert into `table` table}
\usage{
prepare_table_table(df, con)
}
\arguments{
\item{df}{dataframe with table_code and table_name columns}

\item{con}{connection to the database}
}
\value{
a dataframe with the `code`, `name`, `source_id`, `url`, and `notes` columns
for the tables.
}
\description{
Helper function that manually prepares the table table.
Returns table ready to insert into the `table` table with the db_writing family
of functions from UMARimportR using \link[UMARimportR]{sql_function_call}. The input
dataframe must have passed through \link[UMARfetchR]{compute_table_codes} so that
the table_codes and table_names are aligned.
}
