% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{read_codes_from_metadata_excel}
\alias{read_codes_from_metadata_excel}
\title{Little helper function to get the codes out of a metadata file}
\usage{
read_codes_from_metadata_excel(filename)
}
\arguments{
\item{filename}{location of excel file of the `umar_serije_metadata_XX.xlsx` variety}
}
\value{
vector of code names.
}
\description{
does what it says on the tin. no error checking of any note.
}
