% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{split_structure}
\alias{split_structure}
\title{Prepare message describing structure process}
\usage{
split_structure(df)
}
\arguments{
\item{df}{dataframe from structure template}
}
\value{
a list with both the old and the new df
}
\description{
Prepares a log message describing how many new series will be imported
from each worksheet and how many are already there and will be ignored.
}
