% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{update_metadata}
\alias{update_metadata}
\title{Wrapper function for complete metadata pipeline}
\usage{
update_metadata(filename, con, schema, path = "logs/", keep_vintage = FALSE)
}
\arguments{
\item{filename}{location of metadata file}

\item{con}{connection to database}

\item{schema}{name of database schema}

\item{path}{folder to store the log file to}

\item{keep_vintage}{logical value whether or not to keep vintage data}
}
\value{
Nothing, just side effects :). Writes to the database and the excel file and emails logs.
}
\description{
This wrapper function runs the whole pipeline in the \link[UMARfetchR]{main_structure}, while
logging everything to the sink and emailing the logs to the listed recipients
}
