% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{update_structure_excel}
\alias{update_structure_excel}
\title{Update structure table in Excel file}
\usage{
update_structure_excel(filename, df)
}
\arguments{
\item{filename}{name and path of excel file}

\item{df}{updated dataframe.}
}
\value{
nothing,
}
\description{
After running \link[UMARfetchR]{parse_structure} the updated dataframe
with table codes and series codes etc. is written
back into the original Excel file with this funciton
}
