% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{add_new_author}
\alias{add_new_author}
\title{Main function for adding a new author}
\usage{
add_new_author(
  name,
  initials,
  email,
  con,
  schema = "platform",
  data_location = "O:/Avtomatizacija/umar-data",
  overwrite = FALSE
)
}
\arguments{
\item{name}{full name}

\item{initials}{initials, make them unique if already exist}

\item{email}{email}

\item{con}{connection to the database}

\item{schema}{schema name. defaults to "platform"}

\item{data_location}{location where data folders live. defaults to "O://Avtomatizacija//umar-data"}

\item{overwrite}{logical value whether or not to overwrite existing files. defaults to FALSE}
}
\value{
true if completes, otherwise error message
}
\description{
This function inserts the author into the database - checking if the initials are
unique, otherwise you'll need to change them. Then creates the data folder for
that author and places the template excel file in there. And finally adds the
category and cat relationship rows for that author.
}
