% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{check_structure_df}
\alias{check_structure_df}
\title{Check structure data to be parsed}
\usage{
check_structure_df(df, con)
}
\arguments{
\item{df}{dataframe with the structure data, which means at a minumum the columns
source, author, table_name, dimensions, dimension_levels_text, dimension_levels_code,
unit, and interval.}

\item{con}{connection to the database.}
}
\value{
TRUE if passes all checks
}
\description{
This function checks if the input data that is read from an excel file is
in the proper format to be parsed by the parser functions and gives informative
errors if not.
}
