% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_category}
\alias{insert_new_category}
\title{Insert new category}
\usage{
insert_new_category(cat_name, con)
}
\arguments{
\item{cat_name}{character name of category - usually full name of author}

\item{con}{connection to the database.}
}
\value{
nothing
}
\description{
Used to insert the new UMAR categories into the database.
}
