% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{main_data}
\alias{main_data}
\title{Main function for processing data}
\usage{
main_data(filename, codes, con, schema)
}
\arguments{
\item{filename}{path to excel file with timeseries data}

\item{codes}{output of \link[UMARfetchR]{main_structure} i.e. list of series codes}

\item{con}{connection to database}

\item{schema}{schema name}
}
\value{
0 if all sheets are empty, otherwise just message
}
\description{
inputs the excel file with the timeseries data plus a vector of the allowed codes and
imports them into the database.
}
