% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{main_structure}
\alias{main_structure}
\title{Main function for processing structure data}
\usage{
main_structure(filename, con, schema = "platform", keep_vintage = FALSE)
}
\arguments{
\item{filename}{path to excel file with timeseries structure metadata}

\item{con}{connection to database}

\item{schema}{schema name}

\item{keep_vintage}{logical value whether or not to keep old vintages}
}
\value{
if the whole thing goes through without a hitch, returns a character
vector of series codes, otherwise an error will be thrown.
}
\description{
This is the top level funciton to process the structre metadata from a
file an author prepares using the template created with
\link[UMARfetchR]{create_structure_template_excel}. This function:
* reads the structure metadata and checks the data
* computes the columns that are missing
* inserts the data into the database
* updates the original excel file.
}
